Scale to fullscreen using an intermediate texture
The engine now renders everything to a texture target the same size as the native resolution of the game, and then scales that texture to fit the screen.

This is slower than the previous method, which used a camera to achieve the same result, but as everything is scaled at the same time it solves the issues with rounding errors and fixes #5. As we say around here, "muerto el perro se acabó la rabia."